pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "basic_image_downloader"
version = "0.1.0"
description = "Batch downloading of images from single or multiple URLs"
classifiers = [
  "Private :: Do Not Upload",
  "Programming Language :: Python :: 3.11"
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "requests",
  "python-dotenv (>=1.1.1,<2.0.0)",
  "black (>=25.1.0,<26.0.0)",
  "isort (>=6.0.1,<7.0.0)",
]

[project.optional-dependencies]
test = [
  "black",
  "isort[colors]",
]

[project.urls]
Source = "https://github.com/chebishev/BasicImageDownloader"

[tool.black]
exclude = '(\.git|\.mypy_cache|\.venv)'
include = '\.pyi?$'
target-version = ['py311']
line-length = 88

[tool.isort]
profile = "black"